Move org.jabref.logic.database.DuplicateCheck#correlateByWords to org.jabref.logic.util.strings.StringSimilarity. Adapt tests and code accordingly.
