+++ jablib/src/test/java/org/jabref/logic/database/DuplicateCheckTest.java
@@ -2,6 +2,7 @@ package org.jabref.logic.database;
 
 import java.util.stream.Stream;
 
+import org.jabref.logic.util.strings.StringSimilarity;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.BibEntryTypesManager;
@@ -348,9 +349,9 @@ public class DuplicateCheckTest {
         String d2 = "Characterization of Calunus finmarchicus habitat in the North Sea";
         String d3 = "Characterization of Calanus glacialissss habitat in the South Sea";
 
-        assertEquals(1.0, DuplicateCheck.correlateByWords(d1, d2), 0.01);
-        assertEquals(0.78, DuplicateCheck.correlateByWords(d1, d3), 0.01);
-        assertEquals(0.78, DuplicateCheck.correlateByWords(d2, d3), 0.01);
+        assertEquals(1.0, StringSimilarity.correlateByWords(d1, d2), 0.01);
+        assertEquals(0.78, StringSimilarity.correlateByWords(d1, d3), 0.01);
+        assertEquals(0.78, StringSimilarity.correlateByWords(d2, d3), 0.01);
     }
 
     @Test
